+++ toolkit/components/translations/tests/browser/shared-head.js
@@ -761,6 +761,14 @@ function createRecordsForLanguagePair(fromLang, toLang) {
     { fileType: "vocab", name: `vocab.${lang}.spm` },
   ];
 
+  const attachment = {
+    hash: `${crypto.randomUUID()}`,
+    size: `123`,
+    filename: `model.${lang}.intgemm.alphas.bin`,
+    location: `main-workspace/translations-models/${crypto.randomUUID()}.bin`,
+    mimetype: "application/octet-stream",
+  };
+
   if (models.length !== FILES_PER_LANGUAGE_PAIR) {
     throw new Error("Files per language pair was wrong.");
   }
@@ -775,6 +783,7 @@ function createRecordsForLanguagePair(fromLang, toLang) {
       version: TranslationsParent.LANGUAGE_MODEL_MAJOR_VERSION + ".0",
       last_modified: Date.now(),
       schema: Date.now(),
+      attachment,
     });
   }
   return records;